Revert "Update Benefit functionality added"
This reverts commit 2de487c4789ed4b5d3962c9a0e0ddfab91d9f68d.
