+++ src/main/java/com/pla/core/domain/model/BenefitName.java
@@ -33,6 +33,9 @@ public class BenefitName {
 
     public BenefitName(String benefitName) {
         Preconditions.checkNotNull(benefitName);
+        if (benefitName.trim().length() <= 0 || benefitName.length() > 100) {
+            BenefitException.raiseBenefitNameNotProperLengthException();
+        }
         this.benefitName = benefitName;
     }
 }
