+++ src/main/java/com/pla/core/domain/service/BenefitFactory.java
@@ -6,41 +6,34 @@
 
 package com.pla.core.domain.service;
 
-import com.google.common.base.Preconditions;
 import com.pla.core.application.CreateBenefitCommand;
-import com.pla.core.application.UpdateBenefitCommand;
 import com.pla.core.domain.model.Admin;
 import com.pla.core.domain.model.Benefit;
 import com.pla.core.specification.BenefitNameIsUnique;
-import org.nthdimenzion.common.service.JpaRepositoryFactory;
 import org.nthdimenzion.ddd.domain.AbstractDomainFactory;
 import org.nthdimenzion.ddd.domain.annotations.DomainFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.repository.CrudRepository;
 
 /**
  * @author: Samir
  * @since 1.0 09/03/2015
  */
 @DomainFactory
-public class BenefitService extends AbstractDomainFactory {
+public class BenefitFactory extends AbstractDomainFactory {
 
 
     private AdminRoleAdapter adminRoleAdapter;
 
     private BenefitNameIsUnique benefitNameIsUnique;
 
-    private JpaRepositoryFactory jpaRepositoryFactory;
-
     private Logger logger = LoggerFactory.getLogger(AdminRoleAdapter.class);
 
     @Autowired
-    public BenefitService(AdminRoleAdapter adminRoleAdapter, BenefitNameIsUnique benefitNameIsUnique, JpaRepositoryFactory jpaRepositoryFactory) {
+    public BenefitFactory(AdminRoleAdapter adminRoleAdapter, BenefitNameIsUnique benefitNameIsUnique) {
         this.adminRoleAdapter = adminRoleAdapter;
         this.benefitNameIsUnique = benefitNameIsUnique;
-        this.jpaRepositoryFactory = jpaRepositoryFactory;
     }
 
     public Benefit createBenefit(CreateBenefitCommand createBenefitCommand) {
@@ -51,11 +44,5 @@ public class BenefitService extends AbstractDomainFactory {
         Admin admin = adminRoleAdapter.userToAdmin(createBenefitCommand.getUserDetails());
         Benefit benefit = admin.createBenefit(benefitNameIsUnique, benefitId, createBenefitCommand.getBenefitName());
         return benefit;
-    }
-
-    public Benefit updateBenefit(UpdateBenefitCommand updateBenefitCommand) {
-        Preconditions.checkNotNull(updateBenefitCommand.getBenefitId());
-        CrudRepository<Benefit, String> benefitRepository = jpaRepositoryFactory.getCrudRepository(Benefit.class);
-        Benefit benefit = benefitRepository.findOne(updateBenefitCommand.getBenefitId());
     }
 }
